It's like when you're a kid. The first time they tell you that the world's turning, and you just can't quite believe it
'cause everything looks like it's standing still. I can feel it. The turn of the Earth

The ground beneath our feet is spinning at a thousand miles an hour, and the entire planet is hurtling around the sun at sixty-seven thousand miles an hour, and I can feel it. We're falling through space, you and me, clinging to the skin of this tiny little world, and if we let go? That's who I am.